+++ src/insertions.ts
@@ -1,0 +1,9 @@
+export type Subscribable<T> = {
+    subscribe(listner: (value: T) => void): void;
+};
+
+export type Subscription<T> = {
+    next(value: T): void;
+};
+
+export type Instertions = Subscribable<unknown> | Subscription<Event>;
